pyproject.toml
[tool.poetry]
name = "tdd-lab6"
version = "0.1.0"
description = ""
authors = ["Andrii <tap.press.swap@gmail.com>"]
packages = [
    { include = "tdd_lab6", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.2.1"
fastapi = {extras = ["all"], version = "^0.110.1"}
beanie = "^1.25.0"
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
pytest-mock = "^3.14.0"
asgi-lifespan = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
